# Fix/improve available responsibles and assignees performance