--- a/app/models/principal.rb
+++ b/app/models/principal.rb
@@ -42,6 +42,10 @@ class Principal < ActiveRecord::Base
 
   self.table_name = "#{table_name_prefix}users#{table_name_suffix}"
 
+  has_one :preference,
+    dependent: :destroy,
+    class_name: 'UserPreference',
+    foreign_key: 'user_id'
   has_many :members, foreign_key: 'user_id', dependent: :destroy
   has_many :memberships, -> {
     includes(:project, :roles)