--- a/spec/views/api/experimental/projects/show_api_json_spec.rb
+++ b/spec/views/api/experimental/projects/show_api_json_spec.rb
@@ -29,16 +29,15 @@
 require File.expand_path('../../../../../spec_helper', __FILE__)
 
 describe 'api/experimental/projects/show.api.rabl', type: :view do
-  let(:principal) { FactoryGirl.build(:principal) }
-  let(:members)   { FactoryGirl.build_list(:member, 3, principal: principal) }
+  let(:principals) { FactoryGirl.build_list(:principal, 3) }
   let(:types)     { FactoryGirl.build_list(:type,   2) }
 
   let(:project)   {
-    FactoryGirl.build(:project,
-                      possible_responsible_members: members,
-                      possible_assignee_members:    members,
-                      types:                        types
-                     )
+    project = FactoryGirl.build(:project,
+                                possible_responsibles: principals,
+                                possible_assignees:    principals,
+                                types:                 types
+                               )
   }
 
   before do