@TeatroIO 2015-09-23
I've prepared a stage to preview changes. <a href='http://ulferts-fix-improve-available-responsibles-and-assignees-perfor.opf-openproject-2af22bf12c7dfdad4ddf.ttrcloud.com' target='_blank'>Open stage</a> or <a href='http://ulferts-fix-improve-available-responsibles-and-assignees-perfor.opf-openproject-2af22bf12c7dfdad4ddf.ttrcloud.com/teatro.log' target='_blank'>view logs</a>.