@NobodysNightmare 2015-09-23
LGTM, albeit the hack :horse: 

Waiting for travis.