@ulferts 2015-09-24
Thanks for the *ping* @NobodysNightmare. Most seem to be caused by mocks not being adapted. Will take a look.