@NobodysNightmare 2015-09-25
good job on removing that spec duplication while preserving searchability.

Will merge when travis goes green. :horse: